Move options validation to ProviderRegistration.
