+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+* Moved provider options validation to ProviderRegistration
+
 ## [3.19.0-alpha.0] - 2020-06-08
 ### Changed
 * Refactored provider registration, output registration, and controller treatment.
