+++ lib/provider-registration/index.js
@@ -1,10 +1,23 @@
 const chalk = require('chalk')
 const Table = require('easy-table')
 const _ = require('lodash')
+const Joi = require('@hapi/joi')
 const createController = require('./create-controller')
 const createModel = require('./create-model')
 const ProviderRoute = require('./provider-route')
 const ProviderOutputRoute = require('./provider-output-route')
+
+const providerOptionsSchema = Joi.object({
+  cache: Joi.object().keys({
+    retrieve: Joi.function().arity(3).required(),
+    upsert: Joi.function().arity(3).required()
+  }).unknown(true).optional(),
+  routePrefix: Joi.string().optional(),
+  before: Joi.function().arity(2).optional(),
+  after: Joi.function().arity(3).optional(),
+  name: Joi.string().optional(),
+  defaultToOutputRoutes: Joi.boolean().optional()
+}).unknown(true)
 
 module.exports = class ProviderRegistration {
   static create (params) {
@@ -20,6 +33,8 @@ module.exports = class ProviderRegistration {
       koop,
       options = {}
     } = params
+
+    validateOptions(options)
 
     this.options = { ...options, ..._.pick(provider, 'hosts', 'disableIdParam') }
     this.registerOutputRoutesFirst = _.get(options, 'defaultToOutputRoutes', false).toString() === 'true'
@@ -107,4 +122,9 @@ module.exports = class ProviderRegistration {
 
 function getProviderName (provider, options) {
   return options.name || provider.namespace || provider.pluginName || provider.plugin_name || provider.name
+}
+
+function validateOptions (params) {
+  const result = providerOptionsSchema.validate(params)
+  if (result.error) throw new Error(`Provider options ${result.error}`)
 }
